services/api/requirements.txt
pydantic==0.32.2
grpcio==1.24.3
protobuf==3.10.0
fastapi==0.42.0
uvicorn==0.10.3
PyYAML==5.1.2